Refactor to allow module parsers to be defined outside default_utils
